Enabled/added support for different primitive type sizes for int, uint, and float.  - Changed primitive types into an array of primitive types, indexed by    size enumeration type.  - Folded call expression of type(size) into a sym expr of the appropriate    primitive type.  - Chapel-generated C constants are now wrapped with a macro that will    add the appropriate C-style size suffix to the int/uint constant.
Still missing sizes for int(128), uint(128), and complex types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@6157 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/include/expr.h
@@ -181,10 +181,10 @@ class ImportExpr : public Expr {
 
 
 Expr *new_IntLiteral(char *);
-Expr *new_IntLiteral(int i);
+Expr *new_IntLiteral(long long int i);
 Expr *new_UIntLiteral(char *ui_str);
-Expr *new_UIntLiteral(unsigned long u);
-Expr *new_FloatLiteral(char *n, double d);
+Expr *new_UIntLiteral(unsigned long long u);
+Expr *new_FloatLiteral(char *n, long double d);
 Expr *new_StringLiteral(char *str);
 bool get_int(Expr *e, long *i); // false is failure
 bool get_string(Expr *e, char **s); // false is failure
